moved files around and updated eslintrc.js
