+++ .eslintrc.js
@@ -1,4 +1,6 @@
 module.exports = {
   root: true,
   extends: '@react-native-community',
+  rules: {'prettier/prettier': ['error', {endOfLine: 'auto'}]},
+  //Above line added to remove es-lint warnings for cr (lines - aka carrige returns)
 };
