+++ package.json
@@ -14,6 +14,9 @@
     "@react-navigation/bottom-tabs": "^6.5.7",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
+    "@types/react-redux": "^7.1.25",
+    "@types/redux": "^3.6.0",
+    "@types/redux-thunk": "^2.1.0",
     "@types/styled-components": "^5.1.26",
     "formik": "^2.2.9",
     "react": "18.2.0",
@@ -21,6 +24,11 @@
     "react-native": "0.71.7",
     "react-native-safe-area-context": "^4.5.1",
     "react-native-screens": "^3.20.0",
+    "react-native-vector-icons": "^9.2.0",
+    "react-redux": "^8.0.5",
+    "redux": "^4.2.1",
+    "redux-persist": "^6.0.0",
+    "redux-tools": "^3.2.0",
     "styled-components": "^5.3.10",
     "yup": "^1.1.1"
   },
