+++ redux/actions/user.ts
@@ -1,0 +1,8 @@
+/* eslint-disable */
+
+export const setName = () => {
+  return {
+    type: "SET_NAME",
+    payload: name
+  }
+}
