+++ type.d.ts
@@ -1,0 +1,5 @@
+interface IArticle {
+  id: number;
+  title: string;
+  body: string;
+}
